Refactor APIService
